Bash script now for driving the Python script
The previous bash script contained the original bash-driven logic for
moving files. The script was difficult to manage, and it wasn't clear
that it performed correctly all of the time. A Python script was
developed that used the ExifRead module as a pure Python
replacement. That module did not support Fuji RAW files, so a hybrid
solution was developed: use exiftool (via bash) to extract RAW
information, then Python to do the file manipulation based on
that. This script drives that workflow.